requirements.txt
# Core dependencies
requests>=2.31.0
python-dateutil>=2.8.2

# Excel and data processing
openpyxl>=3.1.2
pandas>=2.0.3

# Environment management
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools (optional)
black>=23.7.0
flake8>=6.0.0